feat(admin): rename sidebar entry to Settings + add Movements sub-page
The WODalytics admin surface now has two sub-pages and a clearer label:

  Sidebar section header: "Settings" (was "WODalytics Admin")
    ├── Programs   (existing /admin/programs)
    └── Movements  (new /admin/movements)

The Movements page lifts the pending-movement-review block out of
GymSettings.tsx — that block was already gated by `user.isWodalyticsAdmin`
and didn't really belong on a per-gym settings page. Behavior is
preserved verbatim (load PENDING list, inline edit with parent-search
typeahead, approve / reject), just rendered as a top-level page with
an h1 + EmptyState primitive instead of an in-section h2 + bare
"No pending movements." line.

GymSettings.tsx loses the pending-movement state and handlers
(`pendingMovements`, `editingId`, `mvEditName`, `parentSearch`,
`handleSaveEdit`, `handleReview`, etc.) along with the unused
`useAuth` and `useMovements` imports.

Tests:
- New AdminMovements.test.tsx (renders without crashing, empty state,
  lists pending movements with Edit/Approve/Reject affordances).
- pending-movement-review.spec.ts navigates to /admin/movements and
  waits for the new h1 instead of /settings + the old in-section h2.
- admin-programs.spec.ts updated to assert on the new "Settings"
  section header (was "WODalytics Admin"). Test titles renamed too.

Verified against the live worktree dev stack:
- turbo lint: clean
- web unit: 196 passed
- test:worktree -- api: all suites pass
- test:worktree -- e2e tests/admin-programs.spec.ts
  tests/pending-movement-review.spec.ts: 9/9
- Full E2E suite: 41 passed; same pre-existing programs.spec.ts:277
  ("MEMBER joins a PUBLIC program from Browse") strict-mode flake
  that's been failing on main since before slice 2 — not touched.
