+++ apps/web/tests/admin-programs.spec.ts
@@ -69,15 +69,17 @@ async function teardown(fx: AdminFixture, extraUserId?: string) {
   // user, reused across parallel specs and persistent in the dev DB.
 }
 
-test('admin sees the WODalytics Admin nav and can list + view an unaffiliated program', async ({ page, context }) => {
+test('admin sees the Settings sidebar entry and can list + view an unaffiliated program', async ({ page, context }) => {
   const nameSuffix = randomUUID().slice(0, 8)
   const fx = await seedAdminFixture(nameSuffix)
   try {
     await loginAs(context, fx.adminUserId, 'OWNER')
     await page.goto('/admin/programs')
 
-    // Sidebar nav header for admin.
-    await expect(page.getByText('WODalytics Admin')).toBeVisible()
+    // Sidebar section header for admin (renamed from "WODalytics Admin").
+    // "Gym Settings" still appears in the Staff section, so target the
+    // section-header literal text "Settings" exactly.
+    await expect(page.getByText('Settings', { exact: true })).toBeVisible()
 
     // The seeded program shows up in the list. Match the unique nonce so
     // parallel-running specs that seeded sibling "Admin E2E ..." programs
@@ -138,7 +140,7 @@ test('admin can create a workout under an unaffiliated program (slice 3)', async
   }
 })
 
-test('non-admin user does not see the WODalytics Admin sidebar entry', async ({ page, context }) => {
+test('non-admin user does not see the Settings sidebar entry', async ({ page, context }) => {
   const fx = await seedAdminFixture(randomUUID().slice(0, 8))
   const ts = randomUUID().slice(0, 8)
   const nonAdmin = await prisma.user.create({ data: { email: `admin-e2e-nonadmin-${ts}@test.com` } })
@@ -146,7 +148,8 @@ test('non-admin user does not see the WODalytics Admin sidebar entry', async ({ 
     await loginAs(context, nonAdmin.id, 'MEMBER')
     await page.goto('/feed')
 
-    await expect(page.getByText('WODalytics Admin')).toHaveCount(0)
+    // Sidebar Settings section is admin-only.
+    await expect(page.getByText('Settings', { exact: true })).toHaveCount(0)
 
     // Direct navigation to /admin/programs is allowed (route isn't gated by
     // role on the client; the API enforces the 403). The point of this leg
@@ -157,7 +160,7 @@ test('non-admin user does not see the WODalytics Admin sidebar entry', async ({ 
     await page.goto('/admin/programs')
     await page.waitForURL('**/admin/programs', { timeout: 10000 })
     await expect(page.getByText(/Admin E2E /)).toHaveCount(0)
-    await expect(page.getByText('WODalytics Admin')).toHaveCount(0)
+    await expect(page.getByText('Settings', { exact: true })).toHaveCount(0)
   } finally {
     await teardown(fx, nonAdmin.id)
   }
